Update bitfile with latest .UCF check-in
Forgot to check-in the UCF file so build had unused I2C pads in
random locations. This doesn't affect performance because we also
did the UnusedPin:PullNone trick as well, so electrically this
does nothing. However this commit does allow the blinky LED to
do the thing with I2C.